Enable warning for compiler

+++ src/minhook/src/trampoline.c
@@ -26,7 +26,7 @@
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "intrin_c.h"
+#include <intrin.h>
 #include <windows.h>
 
 #if defined(_M_X64) || defined(__x86_64__)
@@ -197,7 +197,7 @@ BOOL CreateTrampolineFunction(PTRAMPOLINE ct)
                 pCopySrc = &jmp;
                 copySize = sizeof(jmp);
 
-                // Exit the function If it is not in the branch
+                // Exit the function if it is not in the branch
                 finished = (pOldInst >= jmpDest);
             }
         }
@@ -266,7 +266,7 @@ BOOL CreateTrampolineFunction(PTRAMPOLINE ct)
         ct->nIP++;
 
         memcpy((LPBYTE)ct->pTrampoline + newPos, pCopySrc, copySize);
-        newPos += copySize;
+        newPos += (UINT8)copySize;
         oldPos += hs.len;
     }
     while (!finished);
